pr: Create parse pass ustils.
